AMQP-688: Remove Jackson 1 support
JIRA: https://jira.spring.io/browse/AMQP-688

Since Jackson 1 is deprecated for a while it is safe just drop its support in the current major `2.0` version.
