+++ build.gradle
@@ -86,7 +86,6 @@ subprojects { subproject ->
 		commonsCliVersion = '1.2'
 		commonsIoVersion = '2.4'
 		hamcrestVersion = '1.3'
-		jacksonVersion = '1.9.13'
 		jackson2Version = '2.3.2'
 		junitVersion = '4.12'
 		log4jVersion = '2.7'
@@ -199,8 +198,6 @@ project('spring-amqp') {
 		compile ("org.springframework:spring-messaging:$springVersion", optional)
 		compile ("org.springframework:spring-oxm:$springVersion", optional)
 		compile ("org.springframework:spring-context:$springVersion", optional)
-		compile ("org.codehaus.jackson:jackson-core-asl:$jacksonVersion", optional)
-		compile ("org.codehaus.jackson:jackson-mapper-asl:$jacksonVersion", optional)
 		compile ("com.fasterxml.jackson.core:jackson-core:$jackson2Version", optional)
 		compile ("com.fasterxml.jackson.core:jackson-databind:$jackson2Version", optional)
 	}
