--- spring-amqp/src/main/java/org/springframework/amqp/support/converter/JavaTypeMapper.java
@@ -1,37 +1,0 @@
-/*
- * Copyright 2002-2016 the original author or authors.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.springframework.amqp.support.converter;
-
-import org.codehaus.jackson.type.JavaType;
-
-import org.springframework.amqp.core.MessageProperties;
-
-/**
- * Strategy for setting metadata on messages such that one can create the class that needs to be instantiated when
- * receiving a message.
- *
- * @author Mark Pollack
- * @author James Carr
- * @author Sam Nelson
- */
-public interface JavaTypeMapper extends ClassMapper {
-
-   void fromJavaType(JavaType javaType, MessageProperties properties);
-
-   JavaType toJavaType(MessageProperties properties);
-
-}
