+++ spring-amqp/src/test/java/org/springframework/amqp/support/converter/DefaultClassMapperTest.java
@@ -34,7 +34,6 @@ import org.mockito.Spy;
 import org.mockito.runners.MockitoJUnitRunner;
 
 import org.springframework.amqp.core.MessageProperties;
-import org.springframework.amqp.support.converter.JsonMessageConverterTests.Foo;
 
 /**
  * @author James Carr
@@ -140,6 +139,122 @@ public class DefaultClassMapperTest {
 		Map<String, Class<?>> map = new HashMap<String, Class<?>>();
 		map.put(string, class1);
 		return map;
+	}
+
+	public static class Foo {
+
+		private String name = "foo";
+
+		public Foo() {
+		}
+
+		public Foo(String name) {
+			this.name = name;
+		}
+
+		public String getName() {
+			return name;
+		}
+
+		public void setName(String name) {
+			this.name = name;
+		}
+
+		@Override
+		public int hashCode() {
+			final int prime = 31;
+			int result = 1;
+			result = prime * result + ((name == null) ? 0 : name.hashCode());
+			return result;
+		}
+
+		@Override
+		public boolean equals(Object obj) {
+			if (this == obj) {
+				return true;
+			}
+			if (obj == null) {
+				return false;
+			}
+			if (getClass() != obj.getClass()) {
+				return false;
+			}
+			Foo other = (Foo) obj;
+			if (name == null) {
+				if (other.name != null) {
+					return false;
+				}
+			}
+			else if (!name.equals(other.name)) {
+				return false;
+			}
+			return true;
+		}
+
+	}
+
+	public static class Bar {
+
+		private String name = "bar";
+
+		private Foo foo = new Foo();
+
+		public Foo getFoo() {
+			return foo;
+		}
+
+		public void setFoo(Foo foo) {
+			this.foo = foo;
+		}
+
+		public String getName() {
+			return name;
+		}
+
+		public void setName(String name) {
+			this.name = name;
+		}
+
+		@Override
+		public int hashCode() {
+			final int prime = 31;
+			int result = 1;
+			result = prime * result + ((foo == null) ? 0 : foo.hashCode());
+			result = prime * result + ((name == null) ? 0 : name.hashCode());
+			return result;
+		}
+
+		@Override
+		public boolean equals(Object obj) {
+			if (this == obj) {
+				return true;
+			}
+			if (obj == null) {
+				return false;
+			}
+			if (getClass() != obj.getClass()) {
+				return false;
+			}
+			Bar other = (Bar) obj;
+			if (foo == null) {
+				if (other.foo != null) {
+					return false;
+				}
+			}
+			else if (!foo.equals(other.foo)) {
+				return false;
+			}
+			if (name == null) {
+				if (other.name != null) {
+					return false;
+				}
+			}
+			else if (!name.equals(other.name)) {
+				return false;
+			}
+			return true;
+		}
+
 	}
 
 }
