--- spring-amqp/src/test/resources/org/springframework/amqp/support/converter/JsonMessageConverterTests-context.xml
@@ -1,15 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<beans xmlns="http://www.springframework.org/schema/beans"
-	xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-	xsi:schemaLocation="http://www.springframework.org/schema/beans http://www.springframework.org/schema/beans/spring-beans.xsd">
-
-	<bean id="jsonConverterWithDefaultType" class="org.springframework.amqp.support.converter.JsonMessageConverter">
-		<property name="classMapper">
-			<bean class="org.springframework.amqp.support.converter.DefaultClassMapper">
-				<property name="defaultType"
-					value="org.springframework.amqp.support.converter.JsonMessageConverterTests$Foo" />
-			</bean>
-		</property>
-	</bean>
-
-</beans>
